Remove all nullable issues
