+++ src/Microsoft.Tye.Hosting/Watch/CliContext.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace Microsoft.DotNet.Watcher
+{
+    /// <summary>
+    /// This API supports infrastructure and is not intended to be used
+    /// directly from your code. This API may change or be removed in future releases.
+    /// </summary>
+    public static class CliContext
+    {
+        /// <summary>
+        /// dotnet -d|--diagnostics subcommand
+        /// </summary>
+        /// <returns></returns>
+        public static bool IsGlobalVerbose()
+        {
+            bool.TryParse(Environment.GetEnvironmentVariable("DOTNET_CLI_CONTEXT_VERBOSE"), out bool globalVerbose);
+            return globalVerbose;
+        }
+    }
+}
