+++ src/Microsoft.Tye.Hosting/Watch/ProcessSpec.cs
@@ -10,14 +10,15 @@ namespace Microsoft.DotNet.Watcher
 {
     public class ProcessSpec
     {
-        public string Executable { get; set; }
-        public string WorkingDirectory { get; set; }
+        public string? Executable { get; set; }
+        public string? WorkingDirectory { get; set; }
         public IDictionary<string, string> EnvironmentVariables { get; set; } = new Dictionary<string, string>();
-        public IEnumerable<string> Arguments { get; set; }
+        public IEnumerable<string>? Arguments { get; set; }
         public Action<string>? OutputData { get; set; }
         public Action<string>? ErrorData { get; set; }
         public Action<int>? OnStart { get; set; }
-        public string ShortDisplayName()
+        public string? Replica { get; set; }
+        public string? ShortDisplayName()
             => Path.GetFileNameWithoutExtension(Executable);
     }
 }
